organize connectivity extension
